Deal with comments on the pull request: Add function doc string where missing. Invert sense of file times option. Try to explain folder naming option better in doc string and help text. Move folder name function to utils.py
