+++ flickr_download/filename_handlers.py
@@ -7,7 +7,6 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 from collections import defaultdict
-from dateutil import parser
 
 
 def title(pset, photo, suffix):
@@ -21,8 +20,7 @@ def title(pset, photo, suffix):
     """
     if photo.title:
         return '{0}{1}.jpg'.format(photo.title, suffix)
-    else:
-        return idd(pset, photo, suffix)
+    return idd(pset, photo, suffix)
 
 
 def idd(pset, photo, suffix):
@@ -85,13 +83,3 @@ def get_filename_handler(name='title'):
     @return: Function, handler
     """
     return HANDLERS[name]
-
-
-def get_foldername(photo, pattern):
-    """
-    Returns date based foldername
-    """
-    info = photo.getInfo()
-    taken = parser.parse(info['taken'])
-    foldername = taken.strftime(pattern)
-    return foldername
