+++ flickr_download/utils.py
@@ -1,5 +1,7 @@
 import os
 
+from dateutil import parser
+
 
 def sanitize_filepath(fname):
     """
@@ -23,3 +25,16 @@ def get_full_path(pset, photo):
     @return: str, full sanitized path
     """
     return os.path.join(sanitize_filepath(pset), sanitize_filepath(photo))
+
+
+def get_foldername(photo, pattern):
+    """
+    Returns a partial folder path form parsing the photo time with the date format pattern.
+    e.g. The the date Dec 24, 1973 and the pattern '%Y/%m/%d' would return '1973/12/24'
+    @param photo: Flickr.Photo, the photo
+    @param patter: Python date format pattern
+    """
+    info = photo.getInfo()
+    taken = parser.parse(info['taken'])
+    foldername = taken.strftime(pattern)
+    return foldername
